Moved update functions to lib/update.js.
